Move GetFieldLogic to Query/Manipulation namespace
